video-streaming/requirements.txt
opencv-python
paho-mqtt<2.0.0
numpy